完成新增用户

+++ moviesite/moviesite/admin.aspx.cs
@@ -6,6 +6,8 @@ using System.Web.UI;
 using System.Web.UI.WebControls;
 using System.Data;
 using System.Data.SQLite;
+using System.IO;
+using System.Configuration;
 namespace moviesite
 {
     public partial class admin : System.Web.UI.Page
@@ -20,18 +22,67 @@ namespace moviesite
             //{
             //    Response.Redirect("~/index.aspx");
             //}
+            //if (Request.QueryString["logout"] == "true")
+            //{
+            //    Session["username"] = null;
+            //    Session["is_superuser"] = null;
+            //    Response.Redirect("login.aspx");
+            //}
             if (Request.QueryString["delete"]!=null)
             {
-                string sql = string.Format("delete from movie where id=@para1");
+                string sql="";
+                if (Request.QueryString["delete"].ToLower() == "user")
+                {
+                    sql = string.Format("delete from user where id=@para1");
+                }
+                if (Request.QueryString["delete"].ToLower() == "movie")
+                {
+                    sql = string.Format("delete from movie where id=@para1");
+                }
                 SQLiteParameter[] sps = new SQLiteParameter[] {
-                    new SQLiteParameter("@para1",Request.QueryString["delete"])
+                    new SQLiteParameter("@para1",Request.QueryString["id"])
                 };
                 if (SQLiteHelper.ExecuteSql(sql,sps) > 0)
                 {
-                    Response.Redirect("admin.aspx");
+                    Response.Write("<script>alert('删除成功');window.location.href='admin.aspx';</script>");
+                }
+                else
+                {
+                    Response.Write("<script>alert('删除成功');</script>");
                 }
             }
-            
+            if (Request.QueryString["adduser"] != null)
+            {
+                User user = new User();
+                user.UserName = Request.Form["username"];
+                user.PassWord = Request.Form["password"];
+                if (Request.Files.Count > 0)
+                {
+                    user.Avatar =SaveFile(Request.Files["avatar"]);
+                }
+                user.Name = Request.Form["name"];
+                string sql = string.Format("insert into user(username,password,avatar,name) values (@para1,@para2,@para3,@para4)");
+                SQLiteParameter[] sps = new SQLiteParameter[]
+                {
+                        new SQLiteParameter("@para1",user.UserName),
+                        new SQLiteParameter("@para2",user.PassWord),
+                        new SQLiteParameter("@para3",user.Avatar),
+                        new SQLiteParameter("@para4",user.Name)
+                };
+                if (SQLiteHelper.ExecuteSql(sql, sps) > 0)
+                {
+                    Response.Write("<script>alert('新增用户成功！');</script>");
+                    Response.Redirect("admin.aspx");
+                }
+                else
+                {
+                    Response.Write(Request.Url + user.UserName + user.PassWord + user.Avatar + user.Name);
+                }
+            }
+            if (Request.QueryString["advertising"] != null)
+            {
+                
+            }
         }
         public List<Movie> movies
         {
@@ -56,6 +107,46 @@ namespace moviesite
                 movieslist.Add(movie);
             }
             return movieslist;
+        }
+        public List<User> Useres
+        {
+            get
+            {
+                return GetUseres();
+            }
+        }
+        public List<User> GetUseres()
+        {
+            List<User> userlist = new List<User>();
+            string sql = "select * from user";
+            DataSet ds = SQLiteHelper.Query(sql);
+            foreach (DataRow dr in ds.Tables[0].Rows)
+            {
+                User user = new User();
+                user.UserId =Convert.ToInt32(dr["id"]);
+                user.UserName = dr["username"].ToString();
+                user.PassWord = dr["password"].ToString();
+                user.Avatar = dr["avatar"].ToString();
+                user.Description = dr["description"].ToString();
+                user.Name = dr["name"].ToString();
+                user.Email = dr["email"].ToString();
+                user.Qq = dr["qq"].ToString();
+                user.Phone = dr["phone"].ToString();
+                user.Url = dr["url"].ToString();
+                userlist.Add(user);
+            }
+            return userlist;
+        }
+        public string SaveFile(HttpPostedFile f)
+        {
+            string year = DateTime.Now.Year.ToString();
+            string month = DateTime.Now.Month.ToString();
+            string day = DateTime.Now.ToString("dd");
+            string path = string.Format("media\\avatar\\");
+            string filename = string.Format("{0}{1}{2}{3}{4}", year, month, day, DateTime.Now.ToString("mmssffff"), Path.GetExtension(f.FileName));
+            string fullpath = Path.Combine(path, filename);
+            f.SaveAs(Server.MapPath(fullpath));
+            return path + filename;
         }
 
     }
